datatypes of lon and lat changed to double datatype of ids changed to long! order of the waypoints...

+++ src/main/java/ch/zhaw/jroute/model/IWaypointBuilder.java
@@ -5,15 +5,19 @@ import gov.nasa.worldwind.geom.Position;
 import java.util.HashMap;
 import java.util.Observer;
 
-
-
 public interface IWaypointBuilder {
-	
+
 	public void registerObserver(Observer obs);
+
 	public void createWaypointFromPosition(Position waypointPosition);
-	public void setStartWaypoint(WaypointStatusEnum status,Waypoint waypoint);
+
+	public void setStartWaypoint(WaypointStatusEnum status, Waypoint waypoint);
+
 	public void setEndWaypoint(WaypointStatusEnum status, Waypoint waypoint);
-	public HashMap<Float, Waypoint> getWaypointList();
+
+	public HashMap<Long, Waypoint> getWaypointList();
+
 	public Waypoint getStartWaypoint();
+
 	public Waypoint getEndWaypoint();
 }
